Struct google_webfonts1::Webfont

pub struct Webfont {
    pub category: Option<String>,
    pub files: Option<HashMap<String, String>>,
    pub kind: Option<String>,
    pub family: Option<String>,
    pub subsets: Option<Vec<String>>,
    pub last_modified: Option<String>,
    pub version: Option<String>,
    pub variants: Option<Vec<String>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list webfonts (none)

Fields

category

The category of the font.

files

The font files (with all supported scripts) for each one of the available variants, as a key : value map.

kind

This kind represents a webfont object in the webfonts service.

family

The name of the font.

subsets

The scripts supported by the font.

last_modified

The date (format "yyyy-MM-dd") the font was modified for the last time.

version

The font version.

variants

The available variants for the font.